+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- the option `backgroundOpacity` in WallberryTheme to set the brightness of the background.
+
 ## 1.1.0 - 2019-1-21
 ### Changed
 - Updated all module READMEs to make the types of module options more explicit.